Refactor towards language-independence
